requirements.txt
selenium
undetected-chromedriver>=3.5.5
beautifulsoup4
lxml
python-dotenv
setuptools
packaging
